Rename gameplay_keep folder to gameplay_hacker_keep
