+++ assets/objects/gameplay_hacker_keep/bottle_contents.i8.png
