+++ assets/objects/gameplay_hacker_keep/bottle_glass.i8.png
