+++ assets/objects/gameplay_hacker_keep/gameplay_hacker_keep.c
@@ -1,0 +1,5 @@
+#include "config.h"
+
+#if MM_BOTTLE_MODEL
+#include "assets/objects/gameplay_hacker_keep/bottle.c"
+#endif
