+++ assets/objects/gameplay_hacker_keep/gameplay_hacker_keep.h
