--- assets/objects/gameplay_keep/gameplay_hacker_keep.c
@@ -1,5 +1,0 @@
-#include "config.h"
-
-#if MM_BOTTLE_MODEL
-#include "assets/objects/gameplay_keep/bottle.c"
-#endif
